pyproject.toml
[project]
name = "snippet"
version = "0.1.0"
description = ""
authors = [
    {name = "a-pon"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[uvicorn] (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "greenlet (>=3.2.1,<4.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "pytest-asyncio (>=0.26.0,<0.27.0)",
    "python-jose[cryptogrphy] (>=3.4.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "pydantic[email] (>=2.11.3,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
